+++ NEWS.md
@@ -1,3 +1,8 @@
+## v2.0.0 In progress
+### APIs versions
+* Provides `instance-authority-links-statistics v2.0`
+* Removes `linked-bib-update-statistics v1.0`
+
 ## v1.0.0 Released 2023-02-21
 ### APIs versions
 * Provides `instance-authority-links v1.0`